requirements.txt
pycryptodome>=3.15.0
impacket>=0.10.0 